Create PlayerController to attach to the FirstPersonCharacter and make it an origin point of the ray to shoot a gun; delete building ray part from GunController.cs
