+++ Assets/Scripts/PlayerController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: efebd755efe68495da353f0dc18b8bf1
+timeCreated: 1488617485
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
